refactor: UserModal 코드 분리 및 폴더 구조 수정 (#75)
