+++ src/components/ui/Modal/feature/User/User.types.ts
@@ -1,0 +1,13 @@
+export type UserDetail = {
+  id: string
+  name: string
+  email: string
+  gender?: '남성' | '여성' | '기타'
+  nickname?: string
+  birth?: string // yyyy-mm-dd
+  phone?: string
+  role?: string
+  status?: '활성' | '비활성'
+  joinedAt?: string // ISO or display string
+  avatarUrl?: string
+}
